mqttService/requirements.txt
python-dotenv==1.0.1
fastapi==0.115.8
paho-mqtt==2.1.0